refactor common finders logic
