+++ Dockerfile
@@ -7,7 +7,8 @@ FROM python:3-alpine
 
 WORKDIR /proj
 
-RUN apk add -q --no-cache git \
+RUN apk add -q --no-cache \
+    git musl-dev gcc \
     && adduser -S truffleHog
 
 COPY --from=builder dist/truffleHog-3.0.0-py2.py3-none-any.whl /tmp/
