+++ truffleHog/shannon.py
@@ -1,6 +1,6 @@
 import unittest
 import math
-from functools import partial, reduce
+from functools import partial
 
 from toolz.functoolz import compose
 
@@ -92,35 +92,6 @@ class ShannonEntropy:
                 ShannonEntropy.get_strings_of_set, char_set=ShannonEntropy.HEX_CHARS
             ),
         )(word)
-
-
-class HighEntropyStringsFinder:
-    @staticmethod
-    def get_words_entropy(blob):
-        return [
-            {
-                "b64_entropy": ShannonEntropy.find_base64_shannon_entropy(word),
-                "hex_entropy": ShannonEntropy.find_hex_shannon_entropy(word),
-                "word": word,
-            }
-            for word in HighEntropyStringsFinder.get_blob_words(blob)
-        ]
-
-    @staticmethod
-    def get_blob_words(blob):
-        return reduce(lambda x, y: x + y.split(), blob.text, [])
-
-    @staticmethod
-    def filter_words_with_entropy(words_results):
-        return [x for x in words_results if x["b64_entropy"] or x["hex_entropy"]]
-
-    @staticmethod
-    def apply(blob):
-        """given a blob (a thing with a `blob.text : List(str)` field) find the high entropy words"""
-        return compose(
-            HighEntropyStringsFinder.filter_words_with_entropy,
-            HighEntropyStringsFinder.get_words_entropy,
-        )(blob)
 
 
 class TestShannonEntropy(unittest.TestCase):
